+++ lib/src/model/crl/CrlEntryExtensionsData.dart
@@ -1,0 +1,7 @@
+import 'package:basic_utils/src/model/crl/CrlReason.dart';
+
+class CrlEntryExtensionsData {
+  CrlReason? reason;
+
+  CrlEntryExtensionsData({this.reason});
+}