+++ lib/src/model/crl/CrlExtensions.dart
@@ -1,0 +1,9 @@
+class CrlExtensions {
+  String? authorityKeyIdentifier;
+  int? crlNumber;
+
+  CrlExtensions({
+    this.authorityKeyIdentifier,
+    this.crlNumber,
+  });
+}