+++ lib/src/model/crl/RevokedCertificate.dart
@@ -1,0 +1,12 @@
+import 'package:basic_utils/src/model/crl/CrlEntryExtensionsData.dart';
+
+class RevokedCertificate {
+  /// The serialNumber of the certificate
+  BigInt? serialNumber;
+
+  /// The revocation time
+  DateTime? revocationDate;
+
+  /// The extensions
+  CrlEntryExtensionsData? extensions;
+}